requirements.txt
Flask==0.11.1
Flask-Admin
Flask-SQLAlchemy
Flask-API
SQLAlchemy-Searchable==0.10.2
SQLAlchemy-Utils==0.32.9
psycopg2
https://github.com/Songbee/better-bencode/archive/0bb114b9875fac1a6829e0a7bd8ffb933c66c525.tar.gz
requests
smartgetenv
https://github.com/Songbee/beerializer/archive/1b04b82c5ab93f1e32e9f4794997f4ba79920fa0.tar.gz
# beerializer_sqlalchemy==0.1.6
jsonschema
ruamel.yaml

ipython
